requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-openai>=0.2.0
langchain-milvus>=0.1.0
pymilvus[model,milvus_lite]>=2.4.0
sentence-transformers>=3.0.0
faker>=24.0.0
rich>=13.7.0
prompt-toolkit>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
